Refactor: Move the rescheduling-related code to a separate package
