+++ java/code/src/com/suse/manager/maintenance/rescheduling/RescheduleException.java
@@ -12,7 +12,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package com.suse.manager.maintenance;
+package com.suse.manager.maintenance.rescheduling;
 
 
 @SuppressWarnings("serial")
