+++ java/code/src/com/suse/manager/xmlrpc/serializer/RescheduleResultSerializer.java
@@ -19,7 +19,7 @@ import com.redhat.rhn.domain.action.Action;
 import com.redhat.rhn.frontend.xmlrpc.serializer.RhnXmlRpcCustomSerializer;
 import com.redhat.rhn.frontend.xmlrpc.serializer.util.SerializerHelper;
 
-import com.suse.manager.maintenance.RescheduleResult;
+import com.suse.manager.maintenance.rescheduling.RescheduleResult;
 
 import java.io.IOException;
 import java.io.Writer;
